fix(c2): keep L2 pack imports out of answer_engine
lint-imports failed C2 on packs.dms.generative imports inside
answer_engine (pre-existing L2 path). Move generation into
query_service, which is already on the C2 ignore list. No
INVARIANT-CHANGE and no new allowlist entries.
